renamed files from cache to httpdata to be more descriptive
