+++ icann-rdap-client/src/lib.rs
@@ -4,7 +4,7 @@
 use std::{fmt::Display, sync::PoisonError};
 
 use icann_rdap_common::{
-    cache::HttpData,
+    httpdata::HttpData,
     iana::{BootstrapRegistryError, IanaResponseError},
     response::RdapResponseError,
 };
