+++ icann-rdap-client/src/md/types.rs
@@ -2,8 +2,8 @@ use std::any::TypeId;
 
 use lazy_static::lazy_static;
 
-use icann_rdap_common::cache::HttpData;
 use icann_rdap_common::check::string::StringCheck;
+use icann_rdap_common::httpdata::HttpData;
 use icann_rdap_common::response::types::{
     Common, Event, Link, Links, Notices, ObjectCommon, PublicId, Remarks,
 };
