+++ icann-rdap-common/src/httpdata.rs
